requirements.txt
requests>=2.32.4
beautifulsoup4>=4.13.4
pandas
sqlalchemy
python-dotenv
flask
plotly
python-dateutil